requirements.txt
langchain
langchain-openai
streamlit
beautifulsoup4
python-dotenv
chromadb